requirements.txt
fastapi
uvicorn
langchain==0.3.27
langchain_core==0.3.79
langchain-community==0.3.31
dashscope==1.23.9
langgraph==0.6.11
# 以下为过期版本
# langchain-openai
# langchain-wenxin
langchain-deepseek
# langchain-ollama
llama-index==0.13.6
python-dotenv
pydantic==2.11.10
pydantic-settings==2.10.1
numpy
requests
tiktoken
# python-multipart  # 用于FastAPI文件上传支持
httpx             # 用于异步HTTP请求